Displayed peeps in reverse chronological order, refactored spec/helpers modules
