+++ app/controllers/peeps.rb
@@ -1,7 +1,8 @@
-# Controller for peeps route
+# Controller for peeps path
 class Chitter < Sinatra::Base
 
   get '/peeps' do
+    @peeps = Peep.all.reverse
     erb :'/peeps/index'
   end
