+++ app/models/peep.rb
@@ -1,0 +1,6 @@
+class Peep
+  include DataMapper::Resource
+
+  property :id, Serial
+  property :text, String
+end
