+++ app/views/layout.erb
@@ -18,6 +18,11 @@
       width: 250px;
       height: 30px;
     }
+    p {
+      border : 1px solid #00ACED;
+      border-radius: 5px;
+      padding: 5px;
+    }
     .outer_container {
       position: absolute;
       display: table;
