+++ spec/helpers/peep.rb
@@ -1,0 +1,5 @@
+module PeepHelpers
+  def peek_at_peeps
+    visit('/peeps')
+  end
+end
